fixing compiler warning
git-svn-id: https://shawn.svn.sourceforge.net/svnroot/shawn@342 3422d224-1e22-0410-b8a2-c109bc27950b

+++ src/apps/tcpip/storage.h
@@ -105,7 +105,7 @@ public:
 	virtual void writeStorage(tcpip::Storage& store);
 
 	// Some enabled functions of the underlying std::list
-	int size() const { return store.size(); }
+	size_t size() const { return store.size(); }
 
 	StorageType::const_iterator begin() const { return store.begin(); }
 	StorageType::const_iterator end() const { return store.end(); }
